Fix ability to send messages to new conversation in real-time
Addressed issue of broadcasting message from personA to everyone with an open socket (i.e., when intending to send a private message to personB). This was occuring because of using the syntax broadcast.emit, and not using socket.io Rooms. By using the syntax of joining a room (each unique socket based on the id), we can then use the .to method so that new messages will be broadcast to clients that have joined the given room (the socket itself being excluded). In aggregate, this removes messages from being sent to unwanted senders when starting a new conversation.