Fix message order on page load
Addressed issue of message order when returning from db query. Used Sequalize to order by ASC rather than DESC. Then had to change the logic for the convoJSON.latestMessageText so that it would show the most recent message (last in the conversation messages rather than first, as it was when messages were shown in descending order).