Check that sender is authorized to post in specific conversation.
As the code was written, an authenticated user who had specific conversationId would have been allowed to post new messages to that conversation. Hence, I created an authentication method called isConversationParticipant and added to prototypal chain of the Conversation model. If the sender is not authorized to post to a conversation, and we send back a 403 error. Otherwise, they shall proceed as before, and create a new message for said conversation.