Add endpoint to message api and related functionality
Added endpoint to message api to incorporate change in readStatus. An alternate method could have been to add an endpoint for conversations, and then update the messages readStatus in .then() Promise. The /messages/read approach seemed more direct. Update readStatus for messages, based on useEffect within ActiveChat. Currently code is doing a patch request on each re-rendering of ActiveChat, however, may be able to address this with conditional logic, and seeing if there are any existing unread messages.